+++ github/main.py
@@ -1,57 +1,135 @@
-import requests
+import random
+import string
+from time import time
 
-from github.sdk.codexbot_sdk import CodexBot
-from github.sdk.components.server import http_response
+from sdk.codexbot_sdk import CodexBot
+from config import APPLICATION_TOKEN, APPLICATION_NAME, DB, URL, SERVER
 
 
-class GitHub:
-    __name__ = "GitHub"
-
-    routes = []
-
-    sdk = None
-
-    core_host = "http://127.0.0.1:1337/handshake"
+class Github:
 
     def __init__(self):
-        delegated_queue_name = self.handshake()
-        self.sdk = CodexBot(delegated_queue_name)
 
-        self.sdk.set_routes([
-            ('GET', '/test', self.test_route_handler)
+        self.sdk = CodexBot(APPLICATION_NAME, SERVER['host'], SERVER['port'], db_config=DB, token=APPLICATION_TOKEN)
+
+        self.sdk.log("Github module initialized")
+
+        self.sdk.register_commands([
+            ('github_help', 'help', self.help),
+            # ('github_start', 'start', self.start)
         ])
 
-        self.sdk.log("GitHub module inited")
+        self.sdk.set_routes([
+            # ('POST', '/github/{user_token}', self.github_route_handler)
+        ])
+
         self.sdk.start_server()
 
-    def handshake(self):
-        """
-        Requests queue name from core broker
-        :return:
-        """
-        response = requests.post(self.core_host, data={'tool': self.__name__})
+    #
+    #
+    # HELP
+    # todo: move to the class
+    #
 
-        queue_name_received = None
+    async def help(self, payload):
+        self.sdk.log("/help handler fired with payload {}".format(payload))
+        await self.sdk.send_to_chat(
+            payload["chat"],
+            "fuck you"
+        )
 
-        if response.ok and response.status_code == 200:
-            print('Handshake given {}'.format(response.text))
-            queue_name_received = response.text
-        else:
-            Exception("Handshake failed")
+    # #
+    # #
+    # # START
+    # # todo: move to the class
+    # #
+    #
+    # @staticmethod
+    # def generate_user_token():
+    #     return ''.join(random.SystemRandom().choice(string.ascii_uppercase + string.digits) for _ in range(8))
+    #
+    # """
+    # We will store user-to-chat linking in this collection
+    # _id | chat | user | dt_register
+    # """
+    # CHATS_COLLECTION_NAME = 'chats'
+    #
+    # async def start(self, payload):
+    #
+    #     registered_chat = self.sdk.db.find_one(self.CHATS_COLLECTION_NAME, {'chat': payload['chat']})
+    #
+    #     if registered_chat:
+    #         user_token = registered_chat['user']
+    #     else:
+    #         user_token = self.generate_user_token()
+    #         new_chat = {
+    #             'chat': payload['chat'],
+    #             'user': user_token,
+    #             'dt_register': time()
+    #         }
+    #         self.sdk.db.insert(self.CHATS_COLLECTION_NAME, new_chat)
+    #         self.sdk.log("New user registered with token {}".format(user_token))
+    #
+    #     message = "Адрес для отправки уведомлений в этот чат: {}/github/{}\n\n" + \
+    #               "Сообщение отправляйте в POST-параметре «message»"
+    #
+    #     await self.sdk.send_to_chat(
+    #         payload["chat"],
+    #         message.format(URL, user_token)
+    #     )
+    #
+    # #
+    # #
+    # # Route /github/{user_token} handler
+    # #
+    # #
+    # @CodexBot.http_response
+    # async def github_route_handler(self, request):
+    #     """
+    #     :param request: - request data:
+    #         - text
+    #         - post
+    #         - json
+    #         - params - should contain {user_token}
+    #     :return:
+    #     """
+    #     self.sdk.log('Notification accepted {}'.format(request['text']))
+    #
+    #     # Check for route-token passed
+    #     if 'user_token' not in request['params']:
+    #         self.sdk.log("github route handler: user_token is missed")
+    #         return {
+    #             'status': 404
+    #         }
+    #
+    #     # Check for post parameter 'message'
+    #     if 'message' not in request['post'] :
+    #         self.sdk.log("github route handler: param 'message' missed")
+    #         return {
+    #             'text': "Error: param «message» is missed"
+    #         }
+    #
+    #     message = request['post']['message']
+    #     user_token = request['params']['user_token']
+    #
+    #     # Get user data from DB by user token
+    #     registered_chat = self.sdk.db.find_one(self.CHATS_COLLECTION_NAME, {'user': user_token})
+    #
+    #     # Check if chat was registered
+    #     if not registered_chat or 'chat' not in registered_chat:
+    #         self.sdk.log("github route handler: wrong user token passed")
+    #         return {
+    #             'status': 404
+    #         }
+    #
+    #     # Send notification
+    #     await self.sdk.send_to_chat(registered_chat['chat'], message)
+    #
+    #     # Response
+    #     return {
+    #         'text': 'OK!'
+    #     }
 
-        return queue_name_received
 
-
-    @http_response
-    def test_route_handler(self, text, post, json):
-        """
-        Process messages from telegram bot
-        :return:
-        """
-        self.sdk.log("Got test route callback {} {} {}".format(text, post, json))
-
-        return True
-
-
-github = GitHub()
-
+if __name__ == "__main__":
+    github = Github()
