--- notify/config.py
@@ -1,26 +1,0 @@
-##
-# Main URL for services callbacks
-# Should use https protocol
-##
-URL = 'https://57f49d68.ngrok.io'
-
-# HTTP server settings
-SERVER = {
-    'host': '127.0.0.1',
-    'port': 1339
-}
-
-# RabbitMQ server settings
-RABBITMQ = {
-    'host': '127.0.0.1',
-}
-
-# MongoDB
-DB = {
-    'name': 'default',
-    'host': '127.0.0.1',
-    'port': 27017
-}
-
-# Token got from App Manager
-APPLICATION_TOKEN = "8AQFROA7"
